requirements.txt
openai>=1.6.0
python-docx>=1.1.0
matplotlib>=3.7.0
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
PyPDF2>=3.0.0
python-pptx>=0.6.23
streamlit>=1.36.0
Pillow>=10.0.0
pytesseract>=0.3.10
